darshi-main/requirements.txt
fastapi
uvicorn[standard]
gunicorn
python-multipart

# AI
google-genai

# Database (PostgreSQL)
psycopg2-binary>=2.9.9
asyncpg>=0.29.0
sqlalchemy>=2.0.25
geoalchemy2>=0.14.3

# Storage (Cloudflare R2 - S3 compatible)
boto3>=1.34.0

# Core utilities
python-dotenv
pydantic-settings
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
pygeohash
geohash2
httpx[http2]>=0.26.0
python-dateutil>=2.8.2
tenacity>=8.2.3

# OAuth & Authentication
authlib>=1.3.0
itsdangerous>=2.1.2

# Email (Resend)
resend>=0.8.0

# HTTP Client (general use)
requests>=2.31.0

# Security & Rate Limiting
slowapi>=0.1.9
bleach>=6.1.0
redis>=5.0.0

# Push Notifications
pywebpush>=1.14.0

# Image Processing
Pillow>=10.0.0
imagehash>=4.3.1

# Monitoring & Logging
sentry-sdk[fastapi]>=1.40.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
praw>=7.7.1
